pyproject.toml
[project]
name = "noetikon"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[webhooks]>=21.0",
    "anthropic>=0.42.0",
    "httpx[socks]>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "itsdangerous>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "caldav>=1.4.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "httpx>=0.27.0",
    "ruff>=0.9.0",
]

[project.scripts]
noetikon-bot = "noetikon.bot.app:main"
noetikon-admin = "noetikon.admin.app:main"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
